CDAP-9476 fix spark streaming in local mode
Spark2 streaming with checkpointing expects the job jar to be in
the same place in between runs, which means we can never delete
the jar. So in local mode, create the job jar in the app temp dir,
but not in the job specific temp dir.

Also some minor fixes to get preview tests to work.

+++ cdap-spark-core-base/src/main/java/co/cask/cdap/app/runtime/spark/SparkRuntimeService.java
@@ -170,7 +170,7 @@ final class SparkRuntimeService extends AbstractExecutionThreadService {
       initialize();
       SparkRuntimeContextConfig contextConfig = new SparkRuntimeContextConfig(runtimeContext.getConfiguration());
 
-      final File jobJar = generateJobJar(tempDir);
+      final File jobJar = generateJobJar(tempDir, contextConfig.isLocal());
       final List<LocalizeResource> localizeResources = new ArrayList<>();
 
       String metricsConfPath;
@@ -596,8 +596,15 @@ final class SparkRuntimeService extends AbstractExecutionThreadService {
    *
    * @return The generated {@link File} in the given target directory
    */
-  private File generateJobJar(File targetDir) throws IOException {
+  private File generateJobJar(File tempDir, boolean isLocal) throws IOException {
+    // in local mode, Spark Streaming with checkpointing will expect the same job jar to exist
+    // in all runs of the program. This means it can't be created in the temporary directory for the run,
+    // but must persist between runs
+    File targetDir = isLocal ? tempDir.getParentFile() : tempDir;
     File tempFile = new File(targetDir, "emptyJob.jar");
+    if (tempFile.exists()) {
+      return tempFile;
+    }
     JarOutputStream output = new JarOutputStream(new FileOutputStream(tempFile));
     output.close();
     return tempFile;
